컨설팅리스트 조회페이지 추가

+++ src/main/java/egovframework/com/devjitsu/model/consult/TblCnsltDtl.java
@@ -28,7 +28,7 @@ public class TblCnsltDtl {
 
     @Column(name = "CNSLTT_USER_SN", length = 22)
     @Comment("컨설턴트사용자일련번호")
-    private Integer cnslttUserSn;
+    private long cnslttUserSn;
 
     @Column(name = "CNSLTT_DSGN_DT")
     @Comment("컨설턴트지정일")
